moved updateWeekNumber funtion to calenderHelpers file
