+++ src/utils/helpers/calendarHelpers.ts
@@ -151,3 +151,19 @@ export const updateDateInTimeSelections = (newDate, timeIso) => {
 
   return finalDayjs.toISOString()
 }
+
+export const updateWeekNumber = (sundayDate, firstDay, setWeekNumber) => {
+  const secondWeekSunday = dayjs(firstDay).add(7, 'day').format('YYYY-MM-DD')
+  const thirdWeekSunday = dayjs(secondWeekSunday)
+    .add(7, 'day')
+    .format('YYYY-MM-DD')
+  if (sundayDate === firstDay) {
+    setWeekNumber('1')
+  } else if (sundayDate === secondWeekSunday) {
+    setWeekNumber('2')
+  } else if (sundayDate === thirdWeekSunday) {
+    setWeekNumber('3')
+  } else {
+    setWeekNumber('4')
+  }
+}
